Bumped Reactor to latest versions of third-party libs and tweaked a few things to accommodate any minor changes.

+++ reactor-core/src/test/java/reactor/rx/StreamCombinationTests.java
@@ -93,7 +93,7 @@ public class StreamCombinationTests extends AbstractReactorTest {
 		}).consume().debug().toString());
 
 
-		Stream<Integer> stream = Streams.merge(
+		Stream<Long> stream = Streams.merge(
 				Streams.just(1)
 						.map(i -> Streams.range(0, n))
 		);
